requirements.txt
cassandra-driver==3.29.0
fastapi==0.110.0
uvicorn==0.29.0
starlette==0.36.3